pyproject.toml
[project]
name = "automaton-auditor"
version = "2.0.0"
description = "Deep LangGraph Swarm for Autonomous Code Governance"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "astor>=0.8.1",
    "PyPDF2>=3.0.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "python-magic>=0.4.27",
    "tiktoken>=0.5.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
